refactor: simplify knowledge-context.cjs and clean up new test files
- Remove dead `statusTag` variable in `formatAdrLine` (computed then
  immediately overridden by `tag` on the next line).
- Inline the `SUBCOMMANDS` array that existed only to feed the `KNOWN_SUBCOMMANDS`
  Set — eliminated unnecessary intermediate.
- Extract duplicated `loadFile` + `extractSection` helpers from
  `command-adoption.test.ts` into `tests/knowledge/helpers.ts` so the
  knowledge test suite has a single source of truth for these utilities.
- Remove two dead try/catch blocks in the deprecation-notice test that
  captured stderr into variables that were never read before the actual
  capture happened.
